+++ pyoffers/models/__init__.py
@@ -5,8 +5,9 @@ from .core import ModelManager
 from .country import Country, CountryManager
 from .goal import Goal, GoalManager
 from .offer import Offer, OfferManager
+from .raw_log import RawLogManager
 
 
 MODEL_MANAGERS = (
-    AdvertiserManager, ConversionManager, CountryManager, GoalManager, OfferManager
+    AdvertiserManager, ConversionManager, CountryManager, GoalManager, OfferManager, RawLogManager
 )